function that multiplies two integers

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -8,6 +8,6 @@
 */
 int mul(int a, int b)
 {
-int result = a + b;
+long int result = a + b;
 return (result);
 }
